move 40 from naxxramas.h to naxxramas_40.h
this breaks build
